Renamed files
I had to rename files otherwise I would need to change many files within
the project. Initially changing the name of the Owner class was a bad
idea.
