+++ src/main/java/org/springframework/samples/petclinic/owner/Owner.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2002-2013 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.springframework.samples.petclinic.owner;
 
 import java.util.ArrayList;
@@ -6,49 +21,132 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+import javax.validation.constraints.Digits;
+import javax.validation.constraints.NotEmpty;
+
 import org.springframework.beans.support.MutableSortDefinition;
 import org.springframework.beans.support.PropertyComparator;
 import org.springframework.core.style.ToStringCreator;
+import org.springframework.samples.petclinic.model.Person;
 
-public interface Owner {
-	
-	public String getAddress();
+/**
+ * Simple JavaBean domain object representing an owner.
+ *
+ * @author Ken Krebs
+ * @author Juergen Hoeller
+ * @author Sam Brannen
+ * @author Michael Isvy
+ */
+@Entity
+@Table(name = "owners")
+public class Owner extends Person implements OwnerInterface {
+    @Column(name = "address")
+    @NotEmpty
+    private String address;
 
-	public void setAddress(String address) ;
+    @Column(name = "city")
+    @NotEmpty
+    private String city;
 
-	public String getCity();
+    @Column(name = "telephone")
+    @NotEmpty
+    @Digits(fraction = 0, integer = 10)
+    private String telephone;
 
-	public void setCity(String city);
+    @OneToMany(cascade = CascadeType.ALL, mappedBy = "owner")
+    private Set<Pet> pets;
 
-	public String getTelephone();
+    public String getAddress() {
+        return this.address;
+    }
 
-	public void setTelephone(String telephone);
+    public void setAddress(String address) {
+        this.address = address;
+    }
 
-	public Set<Pet> getPetsInternal();
+    public String getCity() {
+        return this.city;
+    }
 
-	public void setPetsInternal(Set<Pet> pets);
+    public void setCity(String city) {
+        this.city = city;
+    }
 
-	public List<Pet> getPets();
+    public String getTelephone() {
+        return this.telephone;
+    }
 
-	public void addPet(Pet pet);
+    public void setTelephone(String telephone) {
+        this.telephone = telephone;
+    }
 
-	/**
-	* Return the Pet with the given name, or null if none found for this Owner.
-	*
-	* @param name to test
-	* @return true if pet name is already in use
-	*/
-	public Pet getPet(String name);
+    public Set<Pet> getPetsInternal() {
+        if (this.pets == null) {
+            this.pets = new HashSet<>();
+        }
+        return this.pets;
+    }
 
-	/**
-	* Return the Pet with the given name, or null if none found for this Owner.
-	*
-	* @param name to test
-	* @return true if pet name is already in use
-	*/
-	public Pet getPet(String name, boolean ignoreNew);
+    public void setPetsInternal(Set<Pet> pets) {
+        this.pets = pets;
+    }
 
-	@Override
-	public String toString();
-	
+    public List<Pet> getPets() {
+        List<Pet> sortedPets = new ArrayList<>(getPetsInternal());
+        PropertyComparator.sort(sortedPets,
+                new MutableSortDefinition("name", true, true));
+        return Collections.unmodifiableList(sortedPets);
+    }
+
+    public void addPet(Pet pet) {
+        if (pet.isNew()) {
+            getPetsInternal().add(pet);
+        }
+        pet.setOwner(this);
+    }
+
+    /**
+     * Return the Pet with the given name, or null if none found for this Owner.
+     *
+     * @param name to test
+     * @return true if pet name is already in use
+     */
+    public Pet getPet(String name) {
+        return getPet(name, false);
+    }
+
+    /**
+     * Return the Pet with the given name, or null if none found for this Owner.
+     *
+     * @param name to test
+     * @return true if pet name is already in use
+     */
+    public Pet getPet(String name, boolean ignoreNew) {
+        name = name.toLowerCase();
+        for (Pet pet : getPetsInternal()) {
+            if (!ignoreNew || !pet.isNew()) {
+                String compName = pet.getName();
+                compName = compName.toLowerCase();
+                if (compName.equals(name)) {
+                    return pet;
+                }
+            }
+        }
+        return null;
+    }
+
+    @Override
+    public String toString() {
+        return new ToStringCreator(this)
+
+                .append("id", this.getId()).append("new", this.isNew())
+                .append("lastName", this.getLastName())
+                .append("firstName", this.getFirstName()).append("address", this.address)
+                .append("city", this.city).append("telephone", this.telephone).toString();
+    }
 }
